LLENADO DE CITA REPOSITORY

+++ src/main/java/com/mediagenda/mediagenda/cita/repository/CitaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mediagenda.mediagenda.cita.model.Cita;
 
-public interface CitaRepository extends JpaRepository<Cita, Long>{ 
+public interface CitaRepository extends JpaRepository<Cita, Integer>{ 
 
     Optional<Cita> findByPacienteRut (String Rut);
